Refactors delete method to utilize classList.contains with the delete button on each card

+++ index.js
@@ -10,12 +10,14 @@ let cardArea = document.querySelector('.card-area');
 let tripPrompt = document.querySelector('#no-trips');
 let intTrips = document.querySelector('#international');
 let domTrips = document.querySelector('#domestic');
+let allBtn = document.querySelector('#view-all');
 
 formInputs.addEventListener('keyup', handleFormInputs);
 btnSave.addEventListener('click', addTrip);
 cardArea.addEventListener('click', deleteCard);
 intTrips.addEventListener('click', returnInternationalArray);
 domTrips.addEventListener('click', returnDomesticArray);
+allBtn.addEventListener('click', displayAllCards);
 window.addEventListener('DOMContentLoaded', handlePageLoad);
 
 function handlePageLoad() {
@@ -25,7 +27,7 @@ function handlePageLoad() {
 }
 
 function handleFormInputs(e) {
-  if (destination.value && startDate.value && endDate.value) {
+  if (destination.value && startDate.value && endDate.value && description.value) {
     btnSave.disabled = false;
   } 
 }
@@ -54,11 +56,14 @@ function addTrip() {
 }
 
 function deleteCard(e) {
-  e.target.closest('.trip-card').remove();
-  var trip = findTrip(e);
+  if (e.target.classList.contains('img-delete')) {
+    e.target.closest('.trip-card').remove();
 
-  trip.deleteFromStorage(tripsArray);
-  displayNoTripsMessage();
+    var trip = findTrip(e);
+  
+    trip.deleteFromStorage(tripsArray);
+    displayNoTripsMessage();
+  }
 }
 
 function findTrip(e) {
@@ -84,18 +89,23 @@ function returnInternationalArray() {
   populateCards(internationalTrips);
 }
 
+function displayAllCards() {
+  cardArea.innerHTML = '';
+  populateCards(tripsArray);
+}
+
 function displayCard(trip) {
   cardArea.insertAdjacentHTML('afterbegin', `<article class="trip-card" data-id="${trip.id}">
         <header>
-          <h3 class="card-destination" contenteditable="true">${trip.destination}</h3>
+          <h3 class="card-destination" >${trip.destination}</h3>
         </header>
         <div class="card-content">
          <div class="card-duration">
-          <p class="card-start-date" contenteditable="true">Start: ${trip.startDate}</p>
-          <p class="card-end-date" contenteditable="true">End: ${trip.endDate}</p>
+          <p class="card-start-date" >Start: ${trip.startDate}</p>
+          <p class="card-end-date" >End: ${trip.endDate}</p>
           </div>
           <p class="card-type">Type: ${trip.type}</p>  
-          <p class="card-body" contenteditable="true">${trip.description}</p>
+          <p class="card-body" >${trip.description}</p>
         </div>
         <footer>
           <img class="img-delete" src="images/delete-button.svg" alt="delete" id="btn-delete">
